+++ project/app/src/androidTest/java/org/owntracks/android/mqtt/MQTTTransitionEventTests.kt
@@ -34,17 +34,6 @@ class MQTTTransitionEventTests :
     TestWithAnMQTTBroker by TestWithAnMQTTBrokerImpl(),
     MockDeviceLocation by GPSMockDeviceLocation() {
 
-//    @DelicateCoroutinesApi
-//    @Before
-//    fun mqttBefore() {
-//        startBroker()
-//    }
-//
-//    @After
-//    fun mqttAfter() {
-//        stopBroker()
-//    }
-
     @After
     fun uninitMockLocation() {
         unInitializeMockLocationProvider()