+++ CHANGELOG.md
@@ -22,6 +22,7 @@
 - An exported log file also contains the threadname for each logentry
 - Certificates are stored in-line in the config, rather than just the names. This should help import/export
 - Removed the undocumented `CLEAR_NOTIFICATIONS` and `REREQUEST_LOCATION_UPDATES` intents
+- Added `clearWaypoints` remote command (#1022)
 
 ### Bug fixes